finish lab2

+++ lab2/DebugExercise/DebugExercise1.java
@@ -6,8 +6,10 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+        //shouldn't be int
+        //int quotient = top / bottom;
+        double quotient = (double)top / bottom;
+        int result = (int) Math.round(quotient);
         return result;
     }
